Refactor to remove exporter code.
