+++ csharp/src/Apache.Arrow.Adbc/Apache.Arrow.Adbc.csproj
@@ -7,9 +7,6 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Apache.Arrow" Version="20.0.0" />
-    <PackageReference Include="OpenTelemetry" Version="1.12.0" />
-    <PackageReference Include="OpenTelemetry.Exporter.Console" Version="1.12.0" />
-    <PackageReference Include="OpenTelemetry.Exporter.OpenTelemetryProtocol" Version="1.12.0" />
     <PackageReference Include="System.Diagnostics.DiagnosticSource" Version="9.0.5" />
     <PackageReference Include="System.Text.Json" Version="8.0.5" />
   </ItemGroup>
@@ -22,5 +19,8 @@
       <PackagePath>\</PackagePath>
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Tracing\FileExporter\" />
   </ItemGroup>
 </Project>
